# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Cliente:
    properties:
      cpf:
        type: string
      data_nascimento:
        type: string
      id:
        type: integer
      nome:
        type: string
    type: object
  model.ClienteCreate:
    properties:
      cpf:
        type: string
      data_nascimento:
        description: Espera-se "YYYY-MM-DD" ou formato RFC3339
        type: string
      nome:
        type: string
    type: object
  model.ClienteWithSaldo:
    properties:
      cpf:
        type: string
      data_nascimento:
        type: string
      id:
        type: integer
      nome:
        type: string
      saldo_devedor:
        type: number
    type: object
...
  types.ErrorResponse:
    properties:
      detail:
        type: string
    type: object
info:
  contact: {}
  description: Aplicação de Banco de Dados para Gerenciamento de Bares
  title: EDNA Bar System
  version: "1.0"
paths:
...
  /clientes:
    get:
      parameters:
      - description: 'Filter by nome using operators: like, ilike, eq, ne. Format:
          operator.value (e.g. like.João)'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by cnpj using operators: eq, ne, like, ilike. Format:
          operator.value (e.g. eq.123456789)'
        in: query
        name: filter-cnpj
        type: string
      - description: 'Sort fields: nome, cnpj. Prefix with ''-'' for desc. Comma separated
          for multiple fields (e.g. -nome,cnpj)'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Cliente'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Clients
      tags:
      - Cliente
    post:
      consumes:
      - application/json
      parameters:
      - description: Cliente payload
        in: body
        name: fornecedor
        required: true
        schema:
          $ref: '#/definitions/model.ClienteCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Cliente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create Cliente
      tags:
      - Cliente
  /clientes/{id}:
    delete:
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Cliente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete Cliente
      tags:
      - Cliente
    get:
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Cliente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Cliente by ID
      tags:
      - Cliente
    put:
      consumes:
      - application/json
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cliente payload
        in: body
        name: fornecedor
        required: true
        schema:
          $ref: '#/definitions/model.ClienteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Cliente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update Cliente
      tags:
      - Cliente
  /clientes/{id}/saldo:
    get:
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ClienteWithSaldo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Fetch Client's Balance
      tags:
      - Cliente
  /clientes/saldos:
    get:
      parameters:
      - description: 'Filter by nome using operators: like, ilike, eq, ne. Format:
          operator.value (e.g. like.João)'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by cnpj using operators: eq, ne, like, ilike. Format:
          operator.value (e.g. eq.123456789)'
        in: query
        name: filter-cnpj
        type: string
      - description: 'Filter by saldo_devedor using operators: eq, ne, gt, lt, gte,
          lte. Format: operator.value (e.g. eq.100)'
        format: float32
        in: query
        name: filter-saldo_devedor
        type: number
      - description: 'Sort fields: nome, cnpj, saldo_devedor. Prefix with ''-'' for
          desc. Comma separated for multiple fields (e.g. -nome,cnpj)'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ClienteWithSaldo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Clients
      tags:
      - Cliente
...
swagger: "2.0"